DeepLabV3Plus-Pytorch-swa/requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
tqdm
matplotlib
visdom
opencv-python-headless
Kornia
